WIP: futher moves traits into distinct sub traits and adds Rust From impls as well
