+++ src/derives/from_argmatches.rs
@@ -9,6 +9,11 @@
 // except according to those terms.
 use proc_macro2;
 use syn;
+use syn::punctuated;
+use syn::token;
+
+use derives;
+use derives::attrs::{Attrs, Kind, Parser, Ty};
 
 pub fn impl_from_argmatches(input: &syn::DeriveInput) -> proc_macro2::TokenStream {
     // use syn::Data::*;
@@ -27,4 +32,99 @@ pub fn impl_from_argmatches(input: &syn::DeriveInput) -> proc_macro2::TokenStrea
 
     // quote!(#inner_impl)
     unimplemented!()
+}
+
+// @TODO impl TryFrom once stable: https://doc.rust-lang.org/std/convert/trait.TryFrom.html
+pub(crate) fn gen_from_argmatches(
+    struct_name: &syn::Ident,
+    fields: &punctuated::Punctuated<syn::Field, token::Comma>,
+) -> proc_macro2::TokenStream {
+    let field_block = gen_constructor(fields);
+
+    quote! {
+        impl ::clap_derive::clap::FromArgMatches for #struct_name {
+            fn from_argmatches(matches: &::clap_derive::clap::ArgMatches) -> Self {
+                #struct_name #field_block
+            }
+        }
+
+        impl From<::clap_derive::clap::ArgMatches> for #struct_name {
+            fn from(&self) -> ::clap_derive::clap::ArgMatches {
+                #struct_name #field_block
+            }
+        }
+    }
+}
+
+pub(crate) fn gen_constructor(
+    fields: &punctuated::Punctuated<syn::Field, token::Comma>,
+) -> proc_macro2::TokenStream {
+    let fields = fields.iter().map(|field| {
+        let attrs = Attrs::from_field(field);
+        let field_name = field.ident.as_ref().unwrap();
+        match attrs.kind() {
+            Kind::Subcommand(ty) => {
+                let subcmd_type = match (ty, derives::sub_type(&field.ty)) {
+                    (Ty::Option, Some(sub_type)) => sub_type,
+                    _ => &field.ty,
+                };
+                let unwrapper = match ty {
+                    Ty::Option => quote!(),
+                    _ => quote!( .unwrap() ),
+                };
+                quote!(#field_name: <#subcmd_type>::from_subcommand(matches.subcommand())#unwrapper)
+            }
+            Kind::FlattenStruct => {
+                quote!(#field_name: ::clap_derive::clap::Clap::from_argmatches(matches))
+            }
+            Kind::Arg(ty) => {
+                use self::Parser::*;
+                let (value_of, values_of, parse) = match *attrs.parser() {
+                    (FromStr, ref f) => (quote!(value_of), quote!(values_of), f.clone()),
+                    (TryFromStr, ref f) => (
+                        quote!(value_of),
+                        quote!(values_of),
+                        quote!(|s| #f(s).unwrap()),
+                    ),
+                    (FromOsStr, ref f) => (quote!(value_of_os), quote!(values_of_os), f.clone()),
+                    (TryFromOsStr, ref f) => (
+                        quote!(value_of_os),
+                        quote!(values_of_os),
+                        quote!(|s| #f(s).unwrap()),
+                    ),
+                    (FromOccurrences, ref f) => (quote!(occurrences_of), quote!(), f.clone()),
+                };
+
+                let occurences = attrs.parser().0 == FromOccurrences;
+                let name = attrs.name();
+                let field_value = match ty {
+                    Ty::Bool => quote!(matches.is_present(#name)),
+                    Ty::Option => quote! {
+                        matches.#value_of(#name)
+                            .as_ref()
+                            .map(#parse)
+                    },
+                    Ty::Vec => quote! {
+                        matches.#values_of(#name)
+                            .map(|v| v.map(#parse).collect())
+                            .unwrap_or_else(Vec::new)
+                    },
+                    Ty::Other if occurences => quote! {
+                        #parse(matches.#value_of(#name))
+                    },
+                    Ty::Other => quote! {
+                        matches.#value_of(#name)
+                            .map(#parse)
+                            .unwrap()
+                    },
+                };
+
+                quote!( #field_name: #field_value )
+            }
+        }
+    });
+
+    quote! {{
+        #( #fields ),*
+    }}
 }
